refactor(app): move the log history viewer state into ainb-app
LogHistoryViewerState, its focus, filter and selection types and the
session summaries move to ainb_app::components::log_history_viewer.

Two ratatui types leave the state:
- session_list_state was a ListState used only for its selected index;
  the renderer already keeps its own ListState in UiState and syncs the
  index each frame. The field becomes selected_session: Option<usize>.
- log_entries_area, the drag-selection hit-test copy, becomes an Area.
