+++ ainb-tui/crates/ainb-app/src/components/mod.rs
@@ -6,6 +6,7 @@ pub mod config_popup;
 pub mod daemons;
 pub mod home_screen_v2;
 pub mod live_logs_stream;
+pub mod log_history_viewer;
 pub mod log_parser;
 pub mod log_reader;
 pub mod log_writer;
